app/requirements.txt
Django
gunicorn
whitenoise
django-import-export
psycopg2-binary
django-tenants
django-tenant-users